Add CRT to run under CP/M.
The Makefile now builds the demo with $100 as its entry point, then it
is linked both against the bios-only crt, and the CP/M-aware crt.
The bios-only version can be run without the operating system, but needs
to be loaded via serial directly using a BIOS rom with a custom loading
routine.
The CP/M-aware version, instead, can be run as a normal executable from
a diskette, but keep in mind that this is a demo, so it is not designed
in such a way that it is safe to go back to CP/M at the end of the demo.
It may be possible, but I won't make any effort to guarantee it.