Run interrupt timer at 300Hz.
300Hz is good because it allows you to run both 50Hz and 60Hz songs with
the same interrupt.
VGM uses a sample rate of 44100Hz (totally unrealistic for the original
hardware on which the SN76489 actually ran, but that's it).
Of course we cannot use the timer at 44100Hz, because that would mean
only 90 Z80 clock cycles in between interrupts, so we use instead a
300Hz interrupt.
One 300Hz interrupt is the same as 147 samples at 44100Hz, which is 5
interrupts (or 147 × 5 = 735 VGM samples) for 60 Hz songs, and 6
interrupts (or 147 x 6 = 882 VGM samples) for 50 Hz songs.